refactor: member, device 도메인 분리
