+++ src/main/java/com/nexters/goalpanzi/application/device/dto/request/UpdatePushActivationStatusCommand.java
@@ -1,4 +1,4 @@
-package com.nexters.goalpanzi.application.member.dto.request;
+package com.nexters.goalpanzi.application.device.dto.request;
 
 public record UpdatePushActivationStatusCommand(
         Long memberId,
