+++ src/main/java/com/nexters/goalpanzi/application/device/event/UpdatePushActivationStatusEvent.java
@@ -1,4 +1,4 @@
-package com.nexters.goalpanzi.application.member.event;
+package com.nexters.goalpanzi.application.device.event;
 
 public record UpdatePushActivationStatusEvent(
         Long memberId,
