+++ src/main/java/com/nexters/goalpanzi/presentation/member/MemberController.java
@@ -3,9 +3,7 @@ package com.nexters.goalpanzi.presentation.member;
 import com.nexters.goalpanzi.application.member.MemberService;
 import com.nexters.goalpanzi.application.member.dto.response.ProfileResponse;
 import com.nexters.goalpanzi.common.argumentresolver.LoginMemberId;
-import com.nexters.goalpanzi.presentation.member.dto.UpdateDeviceTokenRequest;
 import com.nexters.goalpanzi.presentation.member.dto.UpdateProfileRequest;
-import com.nexters.goalpanzi.presentation.member.dto.UpdatePushActivationStatusRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -47,27 +45,5 @@ public class MemberController implements MemberControllerDocs {
         memberService.deleteMember(memberId);
 
         return ResponseEntity.noContent().build();
-    }
-
-    @Override
-    @PatchMapping("/device-token")
-    public ResponseEntity<Void> updateDeviceToken(
-            @LoginMemberId final Long memberId,
-            @RequestBody @Valid final UpdateDeviceTokenRequest request
-    ) {
-        memberService.updateDeviceToken(request.toServiceDto(memberId));
-
-        return ResponseEntity.ok().build();
-    }
-
-    @Override
-    @PatchMapping("/push-activation-status")
-    public ResponseEntity<Void> updatePushActivationStatus(
-            @LoginMemberId final Long memberId,
-            @RequestBody @Valid final UpdatePushActivationStatusRequest request
-    ) {
-        memberService.updatePushActivationStatus(request.toServiceDto(memberId));
-
-        return ResponseEntity.ok().build();
     }
 }
