+++ src/main/java/com/nexters/goalpanzi/presentation/member/MemberControllerDocs.java
@@ -2,9 +2,7 @@ package com.nexters.goalpanzi.presentation.member;
 
 import com.nexters.goalpanzi.application.member.dto.response.ProfileResponse;
 import com.nexters.goalpanzi.common.argumentresolver.LoginMemberId;
-import com.nexters.goalpanzi.presentation.member.dto.UpdateDeviceTokenRequest;
 import com.nexters.goalpanzi.presentation.member.dto.UpdateProfileRequest;
-import com.nexters.goalpanzi.presentation.member.dto.UpdatePushActivationStatusRequest;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
@@ -30,17 +28,5 @@ public interface MemberControllerDocs {
     @Operation(summary = "회원 탈퇴")
     ResponseEntity<Void> deleteMember(
             @Parameter(in = ParameterIn.HEADER, hidden = true) @LoginMemberId final Long memberId
-    );
-
-    @Operation(summary = "디바이스 토큰 갱신")
-    ResponseEntity<Void> updateDeviceToken(
-            @Parameter(in = ParameterIn.HEADER, hidden = true) @LoginMemberId final Long memberId,
-            @RequestBody @Valid final UpdateDeviceTokenRequest request
-    );
-
-    @Operation(summary = "푸시 알림 활성화 상태 변경")
-    ResponseEntity<Void> updatePushActivationStatus(
-            @Parameter(in = ParameterIn.HEADER, hidden = true) @LoginMemberId final Long memberId,
-            @RequestBody @Valid final UpdatePushActivationStatusRequest request
     );
 }
